fix: consider income in recent envelopes
With this, income is considered in the recent envelopes.

In the recent envelopes, income is represented as a recent envelope with the ID `null`.

Therefore, we need to check for this specific case and not use it as suggestion when it is the first
in the list of recent envelopes.

We also need to remove it from the recent envelopes list.