feat!: set default available from to next month
With this, the default AvailableFrom date is set to the first of the month following the transaction, not the month of the transaction.

This makes more sense logically since it enforces funds to only be allocated after they have been received by default.
The value can still be explicitly set if that is required.

The onChange function now just keeps the existing value and does not try to set it to the first of the next month, since that would set a future date every time the field is edited.