fix: available from date and daylight savings time
This fixes a bug where the next month was not correctly calculated during the month on which daylight savings time goes into effect.
As far as I'm aware, that month is March everywhere.

This only affected the "Available From" setting for income transactions for frontend versions 4.0.0 and newer, and only if these transactions were created between midnight and 01:00 your local time.

If you are using the frontend in version 4.0.0 or newer and have created transactions between midnight and 01:00 in March 2025, please verify that the Available From date is correct for those transactions.